requirements/base.txt
Django==3.2.6
python-decouple==3.4
django-environ==0.4.5
psycopg2==2.9.1
django-model-utils==4.1.1
django-cleanup==5.2.0
django-import-export==2.5.0
django-cors-headers==3.8.0
django-allauth==0.45.0
pdfkit==0.6.1
django-floppyforms==1.9.0
django-crispy-forms==1.12.0
django-bootstrap-v5==1.0.5
